Move merged set class to package level
