+++ tools/model-compiler/src/main/java/io/spine/tools/gradle/compiler/DescriptorSetMergerPlugin.java
@@ -21,7 +21,7 @@
 package io.spine.tools.gradle.compiler;
 
 import com.google.common.collect.ImmutableSet;
-import io.spine.code.proto.FileDescriptors;
+import io.spine.code.proto.MergedDescriptorSet;
 import io.spine.tools.gradle.ConfigurationName;
 import io.spine.tools.gradle.SpinePlugin;
 import io.spine.tools.type.MoreKnownTypes;
@@ -96,8 +96,8 @@ public class DescriptorSetMergerPlugin extends SpinePlugin {
             if (descriptorSet.exists()) {
                 files.add(descriptorSet);
             }
-            FileDescriptors.merge(files.build())
-                           .writeTo(descriptorSet);
+            MergedDescriptorSet.create(files.build())
+                               .writeTo(descriptorSet);
 
             // Extend `KnownTypes` with all the type definitions from all the descriptors
             // found in the classpath of the project being built.
